+++ resources/scss/app.scss
@@ -51,6 +51,11 @@
     margin-bottom: 20px;
 }
 
+.card-description{
+    overflow: hidden;
+    height: 100px;
+}
+
 .icon-button{
     background: none;
     padding: 0px;
